+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The Reduce method on a IEnumerable<Range<T>>, returns a new IEnumerable<Range<T>> where all overlapping and touching Ranges have been merged and empty Ranges have been removed
 - The Union method on a IEnumerable<Range<T>> to union two sequences and also reduce the result
 - The Intersect method on a IEnumerable<Range<T>>, returns a new IEnumerable<Range<T>> where only ranges that occur in both sequences are returned
+- IsCompletelyBefore and IsCompletelyBehind methods on Range<T>, the return if a given Range<T> is lower than or greater than the specified range and that they do not touch nor overlap each other
 
 ### Removed 
 - RangeCollection<T> has been removed in favor of Linq extension methods on IEnumerable<Range<T>>